fix(doctor): refuse an unreadable tmux listing instead of reading it as empty
doctor collapsed every `tmux ls` failure into an empty session list — a
permission refusal, a wedged server, a timeout, a missing binary all became
"there are no sessions". That list is a PROTECTIVE set, and an empty one says
nothing is alive:

  - checkOrphanedProcesses builds `live` from it, so every marked process looks
    orphaned and --fix is armed to kill every live session's processes (#2874);
  - liveTmuxHomes builds the "a live tmux session references this home"
    keep-reason from it, guarding an rm -rf;
  - checkRunawayChildren and checkLeakedTmuxSessions rendered as CLEAN for a
    surface af could not read, which is the same fabricated negative in its
    reporting form.

The classifier this needed already existed but was unexported, and doctor's own
comment said so. tmux.ListingFoundNoServer now exports it, beside the
diagnostics it matches: tmux's two definitive no-server answers are an empty
session set, everything else is UNKNOWN. Both #2870 and #2874 got this wrong by
copying, each citing the other as the model, so there is now one place to be
right rather than a shape to re-derive per caller.

Every consumer fails closed and reports the blindness rather than a clean check.

Scope, measured rather than assumed: the temp-home removal is still blocked by
the lock probe on master, so the liveTmuxHomes collapse alone was not shown to
reach the rm -rf. It is fixed regardless, so that removal never depends on the
lock probe being the only surviving guard.

The test stub returned a bare error for `tmux ls`, which is an UNREADABLE
listing, not the definitive no-server answer a box without tmux actually gives.
It now emulates the real diagnostic, so the existing tests exercise "no
sessions" rather than "could not tell" — and a new case covers the difference.

doctor/checks.go crossed the 1000-line limit, so the temp-home liveness rules
moved to their own file: they belong together because every one of them is a
keep-reason guarding the same rm -rf.

Closes #2874
Refs #2905, #2870
